@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,6 +24,7 @@ namespace Seat\Services\Repositories\Character;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
+use Seat\Eveapi\Models\Assets\CharacterAsset;
 use Seat\Eveapi\Models\Character\AssetListContents;
 
 /**
@@ -42,45 +43,38 @@ trait Assets
     public function getCharacterAssets(int $character_id): Collection
     {
 
-        return DB::table('character_asset_lists as a')
+        return CharacterAsset::with('content', 'type')
             ->select(DB::raw('
                 *, CASE
-                when a.locationID BETWEEN 66015148 AND 66015151 then
+                when character_assets.location_id BETWEEN 66015148 AND 66015151 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID=a.locationID-6000000)
-                when a.locationID BETWEEN 66000000 AND 66014933 then
+                      WHERE s.stationID=character_assets.location_id-6000000)
+                when character_assets.location_id BETWEEN 66000000 AND 66014933 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID=a.locationID-6000001)
-                when a.locationID BETWEEN 66014934 AND 67999999 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID=a.locationID-6000000)
-                when a.locationID BETWEEN 60014861 AND 60014928 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID=a.locationID)
-                when a.locationID BETWEEN 60000000 AND 61000000 then
+                      WHERE s.stationID=character_assets.location_id-6000001)
+                when character_assets.location_id BETWEEN 66014934 AND 67999999 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id=character_assets.location_id-6000000)
+                when character_assets.location_id BETWEEN 60014861 AND 60014928 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id=character_assets.location_id)
+                when character_assets.location_id BETWEEN 60000000 AND 61000000 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID=a.locationID)
-                when a.locationID>=61000000 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID=a.locationID)
+                      WHERE s.stationID=character_assets.location_id)
+                when character_assets.location_id BETWEEN 61000000 AND 61001146 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id=character_assets.location_id)
+                when character_assets.location_id > 61001146 then
+                    (SELECT name FROM `universe_structures` AS c
+                     WHERE c.structure_id = character_assets.location_id)
                 else (SELECT m.itemName FROM mapDenormalize AS m
-                    WHERE m.itemID=a.locationID) end
-                    AS location,a.locationId AS locID'))
-            ->selectSub(function ($query) {
-
-                return $query->from('character_asset_list_contents')
-                    ->selectRaw('count(*)')
-                    ->where('parentAssetItemID',
-                        $query->raw('a.itemID'));
-
-            }, 'childContentCount')
-            ->join('invTypes',
-                'a.typeID', '=',
-                'invTypes.typeID')
-            ->join('invGroups',
-                'invTypes.groupID', '=',
-                'invGroups.groupID')
-            ->where('a.characterID', $character_id)
+                    WHERE m.itemID=character_assets.location_id) end
+                AS locationName,
+                character_assets.location_id AS locID'))
+            ->where('character_assets.character_id', $character_id)
             ->get();
     }
 
@@ -96,14 +90,14 @@ trait Assets
      * @return \Illuminate\Support\Collection
      */
     public function getCharacterAssetContents(int $character_id,
-                                                int $parent_asset_id = null,
-                                                int $parent_item_id = null): Collection
+                                              int $parent_asset_id = null,
+                                              int $parent_item_id = null): Collection
     {
 
         $contents = AssetListContents::join('invTypes',
-            'character_asset_list_contents.typeID', '=',
+            'character_assets.type_id', '=',
             'invTypes.typeID')
-            ->where('characterID', $character_id);
+            ->where('character_id', $character_id);
 
         if (! is_null($parent_asset_id))
             $contents = $contents->where('parentAssetItemID', $parent_asset_id);
