@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,7 +24,7 @@ namespace Seat\Services\Repositories\Character;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
-use Seat\Eveapi\Models\Character\ContractItems;
+use Seat\Eveapi\Models\Contracts\ContractItem;
 
 /**
  * Class Contracts.
@@ -39,13 +39,13 @@ trait Contracts
      * @param bool $get
      * @param int  $chunk
      *
-     * @return
+     * @return \Illuminate\Support\Collection
      */
     public function getCharacterContracts(
         int $character_id, bool $get = true, int $chunk = 50)
     {
 
-        $contracts = DB::table(DB::raw('character_contracts as a'))
+        $contracts = DB::table(DB::raw('contract_details as a'))
             ->select(DB::raw(
                 '
                 --
@@ -57,57 +57,64 @@ trait Contracts
                 -- Start Location Lookup
                 --
                 CASE
-                when a.startStationID BETWEEN 66015148 AND 66015151 then
+                when a.start_location_id BETWEEN 66015148 AND 66015151 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID = a.startStationID-6000000)
-                when a.startStationID BETWEEN 66000000 AND 66014933 then
+                      WHERE s.stationID = a.start_location_id-6000000)
+                when a.start_location_id BETWEEN 66000000 AND 66014933 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID = a.startStationID-6000001)
-                when a.startStationID BETWEEN 66014934 AND 67999999 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID = a.startStationID-6000000)
-                when a.startStationID BETWEEN 60014861 AND 60014928 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID = a.startStationID)
-                when a.startStationID BETWEEN 60000000 AND 61000000 then
+                      WHERE s.stationID = a.start_location_id-6000001)
+                when a.start_location_id BETWEEN 66014934 AND 67999999 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id = a.start_location_id-6000000)
+                when a.start_location_id BETWEEN 60014861 AND 60014928 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id = a.start_location_id)
+                when a.start_location_id BETWEEN 60000000 AND 61000000 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID = a.startStationID)
-                when a.startStationID >= 61000000 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID = a.startStationID)
+                      WHERE s.stationID = a.start_location_id)
+                when a.start_location_id >= 61000000 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id = a.start_location_id)
                 else (SELECT m.itemName FROM mapDenormalize AS m
-                    WHERE m.itemID = a.startStationID) end
+                    WHERE m.itemID = a.start_location_id) end
                 AS startlocation,
 
                 --
                 -- End Location Lookup
                 --
                 CASE
-                when a.endstationID BETWEEN 66015148 AND 66015151 then
+                when a.end_location_id BETWEEN 66015148 AND 66015151 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID = a.endStationID-6000000)
-                when a.endStationID BETWEEN 66000000 AND 66014933 then
+                      WHERE s.stationID = a.end_location_id-6000000)
+                when a.end_location_id BETWEEN 66000000 AND 66014933 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID = a.endStationID-6000001)
-                when a.endStationID BETWEEN 66014934 AND 67999999 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID = a.endStationID-6000000)
-                when a.endStationID BETWEEN 60014861 AND 60014928 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID = a.endStationID)
-                when a.endStationID BETWEEN 60000000 AND 61000000 then
+                      WHERE s.stationID = a.end_location_id-6000001)
+                when a.end_location_id BETWEEN 66014934 AND 67999999 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id = a.end_location_id-6000000)
+                when a.end_location_id BETWEEN 60014861 AND 60014928 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id = a.end_location_id)
+                when a.end_location_id BETWEEN 60000000 AND 61000000 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID = a.endStationID)
-                when a.endStationID >= 61000000 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID = a.endStationID)
+                      WHERE s.stationID = a.end_location_id)
+                when a.end_location_id >= 61000000 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id = a.end_location_id)
                 else (SELECT m.itemName FROM mapDenormalize AS m
-                    WHERE m.itemID = a.endStationID) end
+                    WHERE m.itemID = a.end_location_id) end
                 AS endlocation '))
-            ->where('a.characterID', $character_id);
+            ->join('character_contracts', 'character_contracts.contract_id', '=', 'a.contract_id')
+            ->where('character_contracts.character_id', $character_id);
 
         if ($get)
-            return $contracts->orderBy('dateIssued', 'desc')
+            return $contracts->orderBy('date_issued', 'desc')
                 ->paginate($chunk);
 
         return $contracts;
@@ -123,14 +130,15 @@ trait Contracts
     public function getCharacterContractsItems(int $character_id, int $contract_id): Collection
     {
 
-        return ContractItems::join('invTypes',
-            'character_contract_items.typeID', '=',
+        return ContractItem::join('invTypes',
+            'contract_items.type_id', '=',
             'invTypes.typeID')
             ->join('invGroups',
                 'invTypes.groupID', '=',
                 'invGroups.groupID')
-            ->where('characterID', $character_id)
-            ->where('contractID', $contract_id)
+            ->join('character_contracts', 'character_contracts.contract_id', '=', 'contract_items.contract_id')
+            ->where('character_id', $character_id)
+            ->where('character_contracts.contract_id', $contract_id)
             ->get();
 
     }
