@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -43,32 +43,34 @@ trait JumpClone
     {
 
         return DB::table(DB::raw(
-            'character_character_sheet_jump_clones as a'))
+            'character_jump_clones as a'))
             ->select(DB::raw('
                 *, CASE
-                when a.locationID BETWEEN 66015148 AND 66015151 then
+                when a.location_id BETWEEN 66015148 AND 66015151 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID=a.locationID-6000000)
-                when a.locationID BETWEEN 66000000 AND 66014933 then
+                      WHERE s.stationID=a.location_id-6000000)
+                when a.location_id BETWEEN 66000000 AND 66014933 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID=a.locationID-6000001)
-                when a.locationID BETWEEN 66014934 AND 67999999 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID=a.locationID-6000000)
-                when a.locationID BETWEEN 60014861 AND 60014928 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID=a.locationID)
-                when a.locationID BETWEEN 60000000 AND 61000000 then
+                      WHERE s.stationID=a.location_id-6000001)
+                when a.location_id BETWEEN 66014934 AND 67999999 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id=a.location_id-6000000)
+                when a.location_id BETWEEN 60014861 AND 60014928 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id=a.location_id)
+                when a.location_id BETWEEN 60000000 AND 61000000 then
                     (SELECT s.stationName FROM staStations AS s
-                      WHERE s.stationID=a.locationID)
-                when a.locationID>=61000000 then
-                    (SELECT c.stationName FROM `eve_conquerable_station_lists` AS c
-                      WHERE c.stationID=a.locationID)
+                      WHERE s.stationID=a.location_id)
+                when a.location_id>=61000000 then
+                    (SELECT d.name FROM `sovereignty_structures` AS c
+                      JOIN universe_stations d ON c.structure_id = d.station_id
+                      WHERE c.structure_id=a.location_id)
                 else (SELECT m.itemName FROM mapDenormalize AS m
-                    WHERE m.itemID=a.locationID) end
-                    AS location,a.locationId AS locID'))
-            ->join('invTypes', 'a.typeID', '=', 'invTypes.typeID')
-            ->where('a.characterID', $character_id)
+                    WHERE m.itemID=a.location_id) end
+                    AS location,a.location_id AS locID'))
+            ->where('a.character_id', $character_id)
             ->get();
     }
 }
