@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,7 +22,8 @@
 
 namespace Seat\Services\Repositories\Character;
 
-use Seat\Eveapi\Models\Character\MailMessage;
+use Illuminate\Support\Collection;
+use Seat\Eveapi\Models\Mail\MailHeader;
 
 /**
  * Class Mail.
@@ -36,14 +37,14 @@ trait Mail
      *
      * @param int $limit
      *
-     * @return mixed
+     * @return \Illuminate\Support\Collection
      */
-    public function getAllCharacterNewestMail(int $limit = 10)
+    public function getAllCharacterNewestMail(int $limit = 10): Collection
     {
 
         $user = auth()->user();
 
-        $messages = MailMessage::select('messageID', 'senderID', 'characterID', 'senderName', 'title');
+        $messages = MailHeader::select('mail_id', 'from', 'character_id', 'subject');
 
         // If the user is a super user, return all
         if (! $user->hasSuperUser()) {
@@ -63,14 +64,13 @@ trait Mail
                 }
 
                 // Add the collected characterID on previous task to mail records filter
-                $query->whereIn('characterID', $characters)
-                    ->orWhereIn('senderID', $characters);
+                $query->whereIn('character_id', $characters)
+                    ->orWhereIn('from', $characters);
             });
 
         }
 
-        return $messages->groupBy('character_mail_messages.messageID')
-            ->orderBy('character_mail_messages.sentDate', 'desc')
+        return $messages->orderBy('timestamp', 'desc')
             ->limit($limit)
             ->get();
     }
@@ -82,42 +82,37 @@ trait Mail
      * @param bool $get
      * @param int  $chunk
      *
-     * @return
+     * @return \Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Collection
      */
     public function getCharacterMail(
         int $character_id, bool $get = true, int $chunk = 50)
     {
 
-        $mail = MailMessage::join('character_mail_message_bodies',
-            'character_mail_messages.messageID', '=',
-            'character_mail_message_bodies.messageID')
-            ->where('characterID', $character_id);
+        $mail = MailHeader::with('body', 'recipients')
+            ->where('character_id', $character_id);
 
         if ($get)
             return $mail->take($chunk)
-                ->orderBy('sentDate', 'desc')
+                ->orderBy('timestamp', 'desc')
                 ->get();
 
-        return $mail;
+        return $mail->groupBy('mail_id');
     }
 
     /**
-     * Retreive a specific message for a character.
+     * Retrieve a specific message for a character.
      *
      * @param int $character_id
      * @param int $message_id
      *
-     * @return \Seat\Eveapi\Models\Character\MailMessage
+     * @return \Seat\Eveapi\Models\Mail\MailHeader
      */
-    public function getCharacterMailMessage(int $character_id, int $message_id): MailMessage
+    public function getCharacterMailMessage(int $character_id, int $message_id): MailHeader
     {
 
-        return MailMessage::join('character_mail_message_bodies',
-            'character_mail_messages.messageID', '=',
-            'character_mail_message_bodies.messageID')
-            ->where('characterID', $character_id)
-            ->where('character_mail_messages.messageID', $message_id)
-            ->orderBy('sentDate', 'desc')
+        return MailHeader::where('character_id', $character_id)
+            ->where('mail_id', $message_id)
+            ->orderBy('timestamp', 'desc')
             ->first();
     }
 
@@ -140,20 +135,10 @@ trait Mail
         // Get the User for permissions and affiliation
         // checks
         $user = auth()->user();
+        $messages = MailHeader::with('recipients', 'body');
 
-        $messages = MailMessage::join('character_mail_message_bodies',
-            'character_mail_messages.messageID', '=',
-            'character_mail_message_bodies.messageID')
-            ->join(
-                'account_api_key_info_characters',
-                'character_mail_messages.characterID', '=',
-                'account_api_key_info_characters.characterID')
-            ->join(
-                'eve_api_keys',
-                'eve_api_keys.key_id', '=',
-                'account_api_key_info_characters.keyID');
-
-        // If the user is a super user, return all
+        // If a user is not a super user, only return their own mail and those
+        // which they are affiliated to to receive.
         if (! $user->hasSuperUser()) {
 
             $messages = $messages->where(function ($query) use ($user) {
@@ -161,22 +146,20 @@ trait Mail
                 // If the user has any affiliations and can
                 // list those characters, add them
                 if ($user->has('character.mail', false))
-                    $query = $query->whereIn('account_api_key_info_characters.characterID',
-                        array_keys($user->getAffiliationMap()['char']));
+                    $query = $query->whereIn('character_id', array_keys($user->getAffiliationMap()['char']));
 
-                // Add any characters from owner API keys
-                $query->orWhere('eve_api_keys.user_id', $user->id);
+                // Add mail owned by *this* character
+                $query->orWhere('character_id', $user->id);
             });
-
         }
 
         // Filter by messageID if its set
         if (! is_null($message_id))
-            return $messages->where('character_mail_messages.messageID', $message_id)
+            return $messages->where('mail_id', $message_id)
                 ->first();
 
-        return $messages->groupBy('character_mail_messages.messageID')
-            ->orderBy('character_mail_messages.sentDate', 'desc')
+        return $messages->orderBy('timestamp', 'desc')
+            ->groupBy('mail_id')
             ->paginate(25);
     }
 }
