@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,8 +22,8 @@
 
 namespace Seat\Services\Repositories\Character;
 
-use Seat\Eveapi\Models\Character\WalletJournal;
-use Seat\Eveapi\Models\Character\WalletTransaction;
+use Seat\Eveapi\Models\Wallet\CharacterWalletJournal;
+use Seat\Eveapi\Models\Wallet\CharacterWalletTransaction;
 
 /**
  * Class Wallet.
@@ -32,22 +32,20 @@ use Seat\Eveapi\Models\Character\WalletTransaction;
 trait Wallet
 {
     /**
-     * Retreive Wallet Journal Entries for a Character.
+     * Query the eveseat/resources repository for SDE
+     * related information.
      *
      * @param int  $character_id
      * @param bool $get
      * @param int  $chunk
      *
-     * @return
+     * @return mixed
      */
     public function getCharacterWalletJournal(
         int $character_id, bool $get = true, int $chunk = 50)
     {
 
-        $journal = WalletJournal::leftJoin('eve_ref_types',
-            'character_wallet_journals.refTypeID', '=',
-            'eve_ref_types.refTypeID')
-            ->where('characterID', $character_id);
+        $journal = CharacterWalletJournal::where('character_id', $character_id);
 
         if ($get)
             return $journal->orderBy('date', 'desc')
@@ -57,22 +55,22 @@ trait Wallet
     }
 
     /**
-     * Retreive Wallet Transaction Entries for a Character.
+     * Retrieve Wallet Transaction Entries for a Character.
      *
      * @param int  $character_id
      * @param bool $get
      * @param int  $chunk
      *
-     * @return
+     * @return mixed
      */
     public function getCharacterWalletTransactions(
         int $character_id, bool $get = true, int $chunk = 50)
     {
 
-        $transactions = WalletTransaction::where('characterID', $character_id);
+        $transactions = CharacterWalletTransaction::where('character_id', $character_id);
 
         if ($get)
-            return $transactions->orderBy('transactionDateTime', 'desc')
+            return $transactions->orderBy('date', 'desc')
                 ->paginate($chunk);
 
         return $transactions;
