@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -24,6 +24,7 @@ namespace Seat\Services\Repositories\Corporation;
 
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
+use Seat\Eveapi\Models\Wallet\CorporationWalletJournal;
 
 /**
  * Class Ledger.
@@ -32,7 +33,7 @@ use Illuminate\Support\Facades\DB;
 trait Ledger
 {
     /**
-     * Return the Bountry Prize Payout dates for a Corporation.
+     * Return the Bounty Prize Payout dates for a Corporation.
      *
      * @param int $corporation_id
      *
@@ -41,10 +42,9 @@ trait Ledger
     public function getCorporationLedgerBountyPrizeDates(int $corporation_id): Collection
     {
 
-        return DB::table('corporation_wallet_journals')
-            ->select(DB::raw('DISTINCT MONTH(date) as month, YEAR(date) as year'))
-            ->where('corporationID', $corporation_id)
-            ->where('refTypeID', '85')
+        return CorporationWalletJournal::select(DB::raw('DISTINCT MONTH(date) as month, YEAR(date) as year'))
+            ->where('corporation_id', $corporation_id)
+            ->whereIn('ref_type', ['bounty_prizes', 'bounty_prize'])
             ->orderBy('date', 'desc')
             ->get();
     }
@@ -59,14 +59,10 @@ trait Ledger
     public function getCorporationLedgerPIDates(int $corporation_id): Collection
     {
 
-        return DB::table('corporation_wallet_journals')
-            ->select(DB::raw('DISTINCT MONTH(date) as month, YEAR(date) as year'))
-            ->where('corporationID', $corporation_id)
-            ->where(function ($query) {
-
-                $query->where('refTypeID', 96)
-                    ->orWhere('refTypeID', 97);
-            })
+        // TODO : spawn native indexes on month and year inside corporation_wallet_journal table
+        return CorporationWalletJournal::select(DB::raw('DISTINCT MONTH(date) as month, YEAR(date) as year'))
+            ->where('corporation_id', $corporation_id)
+            ->whereIn('ref_type', ['planetary_import_tax', 'planetary_export_tax'])
             ->orderBy('date', 'desc')
             ->get();
     }
@@ -161,17 +157,16 @@ trait Ledger
                                                            int $month = null): Collection
     {
 
-        return DB::table('corporation_wallet_journals')
-            ->select(
-                DB::raw(
-                    'MONTH(date) as month, YEAR(date) as year, ' .
-                    'ROUND(SUM(amount)) as total, ownerName2, ownerID2'
-                ))
-            ->where('corporationID', $corporation_id)
-            ->where('refTypeID', '85')
+        return CorporationWalletJournal::select(
+            DB::raw(
+                'MONTH(date) as month, YEAR(date) as year, ' .
+                'ROUND(SUM(amount)) as total, second_party_id'
+            ))
+            ->where('corporation_id', $corporation_id)
+            ->whereIn('ref_type', ['bounty_prizes', 'bounty_prize'])
             ->where(DB::raw('YEAR(date)'), ! is_null($year) ? $year : date('Y'))
             ->where(DB::raw('MONTH(date)'), ! is_null($month) ? $month : date('m'))
-            ->groupBy('ownerName2')
+            ->groupBy('second_party_id')
             ->orderBy(DB::raw('SUM(amount)'), 'desc')
             ->get();
     }
@@ -190,21 +185,17 @@ trait Ledger
                                                         int $month = null): Collection
     {
 
-        return DB::table('corporation_wallet_journals')
-            ->select(
-                DB::raw(
-                    'MONTH(date) as month, YEAR(date) as year, ' .
-                    'ROUND(SUM(amount)) as total, ownerName1, ownerID1'
-                ))
-            ->where('corporationID', $corporation_id)
+        // TODO : spawn native indexes on month and year inside corporation_wallet_journal table
+        return CorporationWalletJournal::select(
+            DB::raw(
+                'MONTH(date) as month, YEAR(date) as year, ' .
+                'ROUND(SUM(amount)) as total, first_party_id'
+            ))
+            ->where('corporation_id', $corporation_id)
             ->where(DB::raw('YEAR(date)'), ! is_null($year) ? $year : date('Y'))
             ->where(DB::raw('MONTH(date)'), ! is_null($month) ? $month : date('m'))
-            ->where(function ($query) {
-
-                $query->where('refTypeID', 96)
-                    ->orWhere('refTypeID', 97);
-            })
-            ->groupBy('ownerName1')
+            ->whereIn('ref_type', ['planetary_import_tax', 'planetary_export_tax'])
+            ->groupBy('first_party_id')
             ->orderBy(DB::raw('SUM(amount)'), 'desc')
             ->get();
 
