@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,8 +22,7 @@
 
 namespace Seat\Services\Repositories\Corporation;
 
-use Illuminate\Support\Collection;
-use Seat\Eveapi\Models\Corporation\Standing;
+use Seat\Eveapi\Models\Corporation\CorporationStanding;
 
 /**
  * Class Standings.
@@ -34,14 +33,17 @@ trait Standings
     /**
      * Return the standings for a Corporation.
      *
-     * @param $corporation_id
+     * @param int $corporation_id
+     * @param int $chunk
      *
      * @return mixed
      */
-    public function getCorporationStandings(int $corporation_id): Collection
+    public function getCorporationStandings(int $corporation_id, int $chunk = 50)
     {
 
-        return Standing::where('corporationID', $corporation_id)
-            ->get();
+        return CorporationStanding::where('corporation_id', $corporation_id)
+            ->leftJoin('chrFactions', 'from_id', '=', 'factionID')
+            ->orderBy('from_type')
+            ->paginate($chunk);
     }
 }
