@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -23,9 +23,9 @@
 namespace Seat\Services\Repositories\Corporation;
 
 use Illuminate\Support\Collection;
-use Seat\Eveapi\Models\Corporation\CorporationSheetWalletDivision;
-use Seat\Eveapi\Models\Corporation\WalletJournal;
-use Seat\Eveapi\Models\Corporation\WalletTransaction;
+use Seat\Eveapi\Models\Corporation\CorporationDivision;
+use Seat\Eveapi\Models\Wallet\CorporationWalletJournal;
+use Seat\Eveapi\Models\Wallet\CorporationWalletTransaction;
 
 /**
  * Class Wallet.
@@ -43,7 +43,9 @@ trait Wallet
     public function getCorporationWalletDivisions(int $corporation_id): Collection
     {
 
-        return CorporationSheetWalletDivision::where('corporationID', $corporation_id)
+        return CorporationDivision::where('corporation_id', $corporation_id)
+            ->where('type', 'wallet')
+            ->orderBy('division')
             ->get();
     }
 
@@ -57,15 +59,8 @@ trait Wallet
     public function getCorporationWalletDivisionSummary(int $corporation_id): Collection
     {
 
-        return CorporationSheetWalletDivision::join(
-            'corporation_account_balances',
-            'corporation_sheet_wallet_divisions.accountKey', '=',
-            'corporation_account_balances.accountKey')
-            ->select(
-                'corporation_account_balances.balance',
-                'corporation_sheet_wallet_divisions.description')
-            ->where('corporation_account_balances.corporationID', $corporation_id)
-            ->where('corporation_sheet_wallet_divisions.corporationID', $corporation_id)
+        return CorporationDivision::where('corporation_divisions.corporation_id', $corporation_id)
+            ->where('type', 'wallet')
             ->get();
 
     }
@@ -83,13 +78,7 @@ trait Wallet
         int $corporation_id, bool $get = true, int $chunk = 50)
     {
 
-        $journal = WalletJournal::leftJoin(
-            'eve_ref_types', function ($join) {
-
-            $join->on('corporation_wallet_journals.refTypeID', '=',
-                'eve_ref_types.refTypeID');
-        })
-            ->where('corporationID', $corporation_id);
+        $journal = CorporationWalletJournal::where('corporation_id', $corporation_id);
 
         if ($get)
             return $journal->orderBy('date', 'desc')
@@ -106,16 +95,16 @@ trait Wallet
      * @param bool $get
      * @param int  $chunk
      *
-     * @return
+     * @return mixed
      */
     public function getCorporationWalletTransactions(
         int $corporation_id, bool $get = true, int $chunk = 50)
     {
 
-        $transactions = WalletTransaction::where('corporationID', $corporation_id);
+        $transactions = CorporationWalletTransaction::where('corporation_id', $corporation_id);
 
         if ($get)
-            return $transactions->orderBy('transactionDateTime', 'desc')
+            return $transactions->orderBy('date', 'desc')
                 ->paginate($chunk);
 
         return $transactions;
