@@ -3,7 +3,7 @@
 /*
  * This file is part of SeAT
  *
- * Copyright (C) 2015, 2016, 2017  Leon Jacobs
+ * Copyright (C) 2015, 2016, 2017, 2018  Leon Jacobs
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -23,7 +23,6 @@
 namespace Seat\Services\Settings;
 
 use Illuminate\Support\Facades\Cache;
-use Illuminate\Support\Facades\Session;
 use Seat\Services\Exceptions\SettingException;
 
 /**
@@ -62,7 +61,7 @@ abstract class Settings
     protected static $scope = 'global';
 
     /**
-     * Retreive a setting by name.
+     * Retrieve a setting by name.
      *
      * @param      $name
      * @param null $for_id
@@ -73,24 +72,26 @@ abstract class Settings
     public static function get($name, $for_id = null)
     {
 
-        return Cache::rememberForever(
-            self::get_key_prefix($name), function () use ($name, $for_id) {
+        // Pickup the value from the cache if we can. Otherwise,
+        // read the value for the database and cache it for next
+        // time.
+        return Cache::rememberForever(self::get_key_prefix($name,
+            self::get_affected_id($for_id)), function () use ($name, $for_id) {
 
             // Init a new MODEL
             $value = (new static::$model);
 
             // If we are not in the global scope, add a constraint
-            // to be user specific.
+            // to be user group specific.
             if (static::$scope != 'global')
-                $value = $value->where('user_id',
-                    is_null($for_id) ? auth()->user()->id : $for_id);
+                $value = $value->where('group_id', self::get_affected_id($for_id));
 
-            // Retreive the value
+            // Retrieve the value
             $value = $value->where('name', $name)
                 ->value('value');
 
             if ($value)
-                return $value;
+                return json_decode($value);
 
             // If we have no value, check if we can return
             // a default setting
@@ -105,24 +106,49 @@ abstract class Settings
      * Determine the unique prefix for the key by name.
      *
      * @param $name
+     * @param $for_id
      *
      * @return string
      * @throws \Seat\Services\Exceptions\SettingException
      */
-    public static function get_key_prefix($name)
+    public static function get_key_prefix($name, $for_id = null)
     {
 
         // Ensure we have a prefix to work with.
         if (is_null(static::$prefix))
-            throw new SettingException(
-                'No prefix defined. Have you extended and declared $prefix?');
+            throw new SettingException('No prefix defined. Have you extended and declared $prefix?');
 
-        // Prefix user keys with the session_id
+        // Prefix user keys with group_id
         if (static::$scope != 'global')
-            return implode('.', [Session::getId(), static::$prefix, $name]);
+            return implode('.', [$for_id, static::$prefix, $name]);
 
         // Global keys only with the global prefix.
         return implode('.', [static::$prefix, $name]);
+    }
+
+    /**
+     * Determine the effected group id.
+     *
+     * If this is for a specific id use that, otherwise
+     * assume the currently logged in user's id. If we
+     * dont have an already logged in session, well then
+     * we make the $for_id null.
+     *
+     * @param $for_id
+     *
+     * @return int|null
+     */
+    public static function get_affected_id($for_id)
+    {
+
+        // Without auth, return what we have.
+        if (! auth()->check())
+            return $for_id;
+
+        if (is_null($for_id))
+            return auth()->user()->group->id;
+
+        return $for_id;
     }
 
     /**
@@ -139,14 +165,16 @@ abstract class Settings
         $db = (new static::$model);
 
         // If we are not in the global scope, add a constraint
-        // to be user specific.
+        // to be user group specific.
         if (static::$scope != 'global')
-            $db = $db->where('user_id',
-                is_null($for_id) ? auth()->user()->id : $for_id);
+            $db = $db->where('group_id', self::get_affected_id($for_id));
 
-        // Retreive the value
+        // Retrieve the value
         $db = $db->where('name', $name)
             ->first();
+
+        // By default, json encode values.
+        $encoded_value = json_encode($value, JSON_NUMERIC_CHECK);
 
         // Check if we have a value, else create a new
         // instance
@@ -155,18 +183,17 @@ abstract class Settings
 
         $db->fill([
             'name'  => $name,
-            'value' => $value,
+            'value' => $encoded_value,
         ]);
 
         // Again, if we are not in the global context, then
         // we need to constrain this setting to a user.
         if (static::$scope != 'global')
-            $db->user_id = is_null($for_id) ? auth()->user()->id : $for_id;
+            $db->group_id = self::get_affected_id($for_id);
 
         $db->save();
 
         // Update the cached entry with the new value
-        Cache::forever(self::get_key_prefix($name), $value);
-
+        Cache::forever(self::get_key_prefix($name, self::get_affected_id($for_id)), $value);
     }
 }
